feat(examples): move all schema to common location
