+++ fetchai/communication.py
@@ -1,67 +1,13 @@
-import base64
-import hashlib
 import json
-import struct
-from typing import Optional, Any, Callable, Self
+from typing import Optional, Any, Callable
 from uuid import uuid4
 
 import requests
-from pydantic import BaseModel, UUID4
 
 from fetchai.crypto import Identity
 from fetchai.logger import logger
 from fetchai.registration import DEFAULT_ALMANAC_API_URL
-
-JsonStr = str
-
-
-class Envelope(BaseModel):
-    version: int
-    sender: str
-    target: str
-    session: UUID4
-    schema_digest: str
-    protocol_digest: Optional[str] = (
-        "proto:a03398ea81d7aaaf67e72940937676eae0d019f8e1d8b5efbadfef9fd2e98bb2",
-    )
-    payload: Optional[str] = None
-    expires: Optional[int] = None
-    nonce: Optional[int] = None
-    signature: Optional[str] = None
-
-    def encode_payload(self, value: JsonStr):
-        self.payload = base64.b64encode(value.encode()).decode()
-
-    def decode_payload(self) -> str:
-        if self.payload is None:
-            return ""
-
-        return base64.b64decode(self.payload).decode()
-
-    def sign(self, identity: Identity):
-        try:
-            self.signature = identity.sign_digest(self._digest())
-        except Exception as err:
-            raise ValueError(f"Failed to sign envelope: {err}") from err
-
-    def verify(self) -> bool:
-        if self.signature is None:
-            raise ValueError("Envelope signature is missing")
-        return Identity.verify_digest(self.sender, self._digest(), self.signature)
-
-    def _digest(self) -> bytes:
-        hasher = hashlib.sha256()
-        hasher.update(self.sender.encode())
-        hasher.update(self.target.encode())
-        hasher.update(str(self.session).encode())
-        hasher.update(self.schema_digest.encode())
-        if self.payload is not None:
-            hasher.update(self.payload.encode())
-        if self.expires is not None:
-            hasher.update(struct.pack(">Q", self.expires))
-        if self.nonce is not None:
-            hasher.update(struct.pack(">Q", self.nonce))
-        return hasher.digest()
+from fetchai.schema import JsonStr, Envelope, AgentMessage
 
 
 def lookup_endpoint_for_agent(agent_address: str) -> str:
@@ -137,24 +83,6 @@ def send_message_to_agent(
     )
     r.raise_for_status()
     logger.info("Sent message to agent", extra=request_meta)
-
-
-class AgentMessage(BaseModel):
-    # The address of the sender of the message.
-    sender: str
-    # The address of the target of the message.
-    target: str
-    # The payload of the message.
-    payload: Any
-
-    @classmethod
-    def from_envelope(cls, envelope: Envelope) -> Self:
-        json_payload = envelope.decode_payload()
-        payload = json.loads(json_payload)
-
-        return AgentMessage(
-            sender=envelope.sender, target=envelope.target, payload=payload
-        )
 
 
 def parse_message_from_agent(content: JsonStr) -> AgentMessage:
