+++ fetchai/tests/test_communication.py
@@ -4,6 +4,7 @@ import uuid
 import pytest
 from requests_mock import Mocker as RequestsMocker
 
+import fetchai.schema
 from fetchai import communication
 from fetchai.crypto import Identity
 
@@ -41,7 +42,7 @@ class TestSendMessageToAgent:
         assert mock.last_request.port == 80
         assert mock.last_request.path == f"/fake_endpoint/{agent_name}"
 
-        payload = communication.Envelope.model_validate_json(mock.last_request.text)
+        payload = fetchai.schema.Envelope.model_validate_json(mock.last_request.text)
         assert (
             payload.protocol_digest
             == "proto:a03398ea81d7aaaf67e72940937676eae0d019f8e1d8b5efbadfef9fd2e98bb2"
@@ -60,11 +61,11 @@ class TestParseMessageFromAgent:
         return {"hello": "there"}
 
     @pytest.fixture
-    def envelope(self, identity: Identity, payload: dict) -> communication.Envelope:
+    def envelope(self, identity: Identity, payload: dict) -> fetchai.schema.Envelope:
         session_id = uuid.uuid4()
         json_payload = json.dumps(payload)
 
-        content = communication.Envelope(
+        content = fetchai.schema.Envelope(
             version=1,
             sender=identity.address,
             target="agent_two",
@@ -77,7 +78,7 @@ class TestParseMessageFromAgent:
 
         return content
 
-    def test_happy_path(self, envelope: communication.Envelope, payload: dict):
+    def test_happy_path(self, envelope: fetchai.schema.Envelope, payload: dict):
         agent_message = communication.parse_message_from_agent(
             envelope.model_dump_json()
         )
